fix: merge tests to ensure sequence
